+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 7.76.0 <span class="changelog-date">(2023-07-09)</span>
+
 * Fixed Fashion Loadouts being unable to store an Ornament for the Titan exotic Loreley Splendor Helm.
 
 ## 7.75.0 <span class="changelog-date">(2023-07-02)</span>